pyproject.toml
[project]
name = "tasktray"
version = "0.2.0"
description = "Quick menu in the system tray easy configuration with yaml file"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.3.0",
    "pystray>=0.19.5",
    "pyyaml>=6.0.3",
    "pywin32; sys_platform == 'win32'",
    "pyside6>=6.9.3",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
]
